Enhancement Some Code Lines

+++ repository/search/src/test/kotlin/com/repository/search/repository/CategoriesRepositoryImplTest.kt
@@ -7,8 +7,8 @@ import com.repository.search.dataSource.remote.GenresRemoteDataSource
 import com.repository.search.dto.GenreDto
 import com.repository.search.dto.GenresDto
 import com.repository.search.entity.GenreEntity
-import com.repository.search.exception.NoCategoriesFoundException
-import com.repository.search.exception.NoInternetConnectionException
+import com.domain.search.exception.NoCategoriesFoundException
+import com.domain.search.exception.NoInternetConnectionException
 import io.mockk.*
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.test.runTest
@@ -43,7 +43,7 @@ class CategoriesRepositoryImplTest {
         val result = repository.getAllCategories()
 
         // Then
-        assertEquals(localGenres.map { CategoryModel(it.id.toInt(), it.name) }, result)
+        assertEquals(localGenres.map { CategoryModel(it.id, it.name) }, result)
         coVerify(exactly = 0) { genresRemoteDataSource.getAllGenres() }
     }
